Fix errors thrown by the Frankenstein GameObjectInspector
To implement docstrings for objects, a custom inspector for the
GameObject class was created.
To keep the original GUI while making additions, an improperly
initialized instance of the original inspector was used, which caused
lots of errors.

Instead, we now keep the original GameObject inspector, but inject our
custom drawing code at the end of one of its methods at runtime using
the Harmony2 library.