Fix: Camera sometimes did not capture all players when zooming out
Removing the Soft and Dead zone of the Cinemachine MultiTargetCamera
seemed to do the trick.
Looks like those settings and the FramingTransposer for TargetGroups are
not compatible.

Since we always had problems with Cinemachine, we should consider
writing our own camera controller, if we find during testing that this
still doesnt work correctly or other bugs arise.