Fix: Boss respawned when room was entered again
Reason: Race condition between loading the saved state of switch (Enemy spawner)
and the BossSpawner behaviour checkking for the state

Fixed by loading saved state of switch earlier in Awake phase.

Changing the behaviour of Switches is a bit dangerous since it can break
a lot of things. We will have to see in the tests, whether I broke
anything with this change